backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
google-cloud-firestore==2.14.0
python-dotenv==1.0.0
haversine==2.8.0
